chore: release beta version 0.1.1 (#1)
* chore: add base structure
* feat: add base url validation
* feat: add authorization and basic client options
* chore: prep initial beta release
* ci: fix broken link to Rye setup
rye is a subdomain of astral
* ci: fix test script
* fix: solve annotation issues in type checking
In Python <3.11, Python throws a TypeError indicating that Or operand is not supported...
* refactor: switch from public to private modules
* feat: add spark uri locator
* feat: add spark service execute
* feat: add oauth2 resource for retrieving access token
This also supports retries when access token is missing or expired.
* ci: fix linters
* feat: add more verbs for Spark services
The `get_schema`, `get_metadata`, and `get_versions` verbs retrieve the schema, metadata, and versions of a specified service. These methods enhance the functionality of the Spark services module.
* feat: add batches module for executing sync batch
* docs: add basic guidelines
* chore: update version to 0.1.0-beta.5
These changes allow the execute() method to support both single and multiple inputs. This sync batch execute is longer supported within Batches. Batches API will be dedicated to only async batch operations.
* feat: add support for batch api
* docs(wip): add guidelines for batch api
* chore: prep pre-release stable version 0.1.0
* chore: migrate to coherent account